report-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
reportlab==4.2.5
redis==5.0.7
python-jose==3.3.0
python-dotenv==1.0.1
pymongo==4.8.0

